+++ apps/web/src/entities/alumni-contacts/config/alumniContactsAdmissionYearFilter.ts
@@ -1,0 +1,4 @@
+export const ALUMNI_CONTACTS_ADMISSION_YEAR_FILTER = {
+  MIN: 1972,
+  MAX: new Date().getFullYear(),
+} as const;
