+++ apps/web/src/entities/alumni-contacts/config/index.ts
@@ -6,3 +6,4 @@ export {
   ALUMNI_CONTACTS_ACADEMIC_STATUS_FILTER_OPTION,
   type AlumniContactsAcademicStatusFilterOption,
 } from './alumniContactsAcademicStatusFilterOption';
+export { ALUMNI_CONTACTS_ADMISSION_YEAR_FILTER } from './alumniContactsAdmissionYearFilter';
