+++ apps/web/src/entities/alumni-contacts/index.ts
@@ -3,6 +3,7 @@ export {
   type AlumniContactsSortFilterOption,
   ALUMNI_CONTACTS_ACADEMIC_STATUS_FILTER_OPTION,
   type AlumniContactsAcademicStatusFilterOption,
+  ALUMNI_CONTACTS_ADMISSION_YEAR_FILTER,
 } from './config';
 export {
   useAlumniContactsFilterStore,
