+++ apps/web/src/features/alumni-contacts/model/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useAlumniContactsSearchInput } from './useAlumniContactsSearchInput';
+export { useAlumniContactsAcademicStatusFilterButton } from './useAlumniContactsAcademicStatusFilterButton';
