+++ apps/web/src/widgets/alumni-contacts/model/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useAlumniContactsFilterGroup } from './useAlumniContactsFilterGroup';
+export { useAlumniContactsAcademicFilterSheetModal } from './useAlumniContactsAcademicFilterSheetModal';
+export { useAlumniContactsAdmissionYearFilterSelectGroup } from './useAlumniContactsAdmissionYearFilterSelectGroup';
