+++ apps/web/src/widgets/alumni-contacts/model/index.ts
@@ -1,0 +1,5 @@
+export {
+  useAlumniContactsFilterGroup,
+  useAlumniContactsAcademicFilterSheetModal,
+  useAlumniContactsAdmissionYearFilterSelectGroup,
+} from './hooks';
